Removed a Google build file reference for FLAGS.
Tests are not working for tcpdump as someone thought it would be ok to
change the output format in tcpdump 4.1.1.  This was created using
3.6.2.  We could adjust the parsing to match the new format, but that
seems like a temporary solution.  Not sure what the right answer is, but
it may involve using a native pcap solution to read the binary files.